+++ apps/server/src/services/app-service.test.ts
@@ -533,14 +533,6 @@ describe.sequential("app service", () => {
             transactionsInitialDays: 45,
             automaticSyncConcurrency: 2
           },
-          SALT_EDGE: {
-            environment: "test",
-            appId: "",
-            secret: "",
-            consentDays: 90,
-            transactionsFetchDays: 90,
-            automaticSyncConcurrency: 2
-          }
         })
       } as never,
       runtime: {
@@ -654,14 +646,6 @@ describe.sequential("app service", () => {
             transactionsInitialDays: 45,
             automaticSyncConcurrency: 1
           },
-          SALT_EDGE: {
-            environment: "test",
-            appId: "",
-            secret: "",
-            consentDays: 90,
-            transactionsFetchDays: 90,
-            automaticSyncConcurrency: 2
-          }
         })
       } as never
     });